+++ src/librssguard/gui/webviewers/webengine/webengineviewer.cpp
@@ -1,6 +1,6 @@
 // For license of this file, see <project-root-folder>/LICENSE.md.
 
-#include "gui/webengine/webengineviewer.h"
+#include "gui/webviewers/webengine/webengineviewer.h"
 
 #include "definitions/definitions.h"
 #include "gui/dialogs/formmain.h"
