+++ src/librssguard/gui/webviewers/webviewer.h
