+++ src/librssguard/network-web/webengine/webenginepage.cpp
@@ -3,7 +3,7 @@
 #include "network-web/webengine/webenginepage.h"
 
 #include "definitions/definitions.h"
-#include "gui/webengine/webengineviewer.h"
+#include "gui/webviewers/webengine/webengineviewer.h"
 #include "miscellaneous/application.h"
 #include "network-web/adblock/adblockmanager.h"
 #include "network-web/adblock/adblockrequestinfo.h"
